Refactor code, remove/add comments, change imports
