+++ src/db/models/conversation.model.ts
@@ -1,7 +1,7 @@
 import { Schema, Document, model, Types } from 'mongoose';
-import { MessageDocument } from 'db/models/message.model';
-import { ConversationStatus } from 'db/enum/enums';
-import { EvaluationDocument } from 'db/models/evaluation.model';
+import { MessageDocument } from '@db/models/message.model';
+import { ConversationStatus } from '@enums/conversation-status.enum';
+import { EvaluationDocument } from '@db/models/evaluation.model';
 
 interface ConversationDocument extends Document {
   messages: Types.ObjectId[] | MessageDocument[];
