+++ src/db/models/message.model.ts
@@ -1,12 +1,12 @@
-import { MsgTypes } from 'db/enum/enums';
+import { MsgType } from '@enums/msg-type.enum';
 import { Schema, Document, model } from 'mongoose';
-import { MsgSender } from 'db/enum/enums';
+import { MsgSender } from '@enums/msg-sender.enum';
 import { BaseMessage } from 'langchain/schema';
 
 interface MessageDocument extends Document {
   sender: MsgSender;
   text: string;
-  type: MsgTypes;
+  type: MsgType;
   timestamp: Date;
   intermediateMsg: string | null;
   action: string | null;
@@ -22,7 +22,7 @@ const messageSchema: Schema = new Schema(
   {
     sender: { type: String, enum: Object.values(MsgSender), required: true },
     text: { type: String, required: true },
-    type: { type: String, enum: Object.values(MsgTypes), required: true },
+    type: { type: String, enum: Object.values(MsgType), required: true },
     timestamp: { type: Date, required: true },
     intermediateMsg: { type: String },
     action: { type: String },
