+++ src/db/repositories/agent.repository.ts
@@ -1,6 +1,6 @@
 import { Model } from 'mongoose';
-import { BaseRepository } from 'db/repositories/base.repository';
-import { AgentDocument } from 'db/models/agent.model';
+import { BaseRepository } from '@db/repositories/base.repository';
+import { AgentDocument } from '@db/models/agent.model';
 
 class AgentRepository extends BaseRepository<AgentDocument> {
   constructor(model: Model<AgentDocument>) {
