+++ src/enums/evaluation-status.enum.ts
@@ -1,0 +1,7 @@
+enum EvaluationStatus {
+  EVALUATED = 'evaluated',
+  IN_PROGRESS = 'in_progress',
+  NOT_EVALUATED = 'not_evaluated',
+}
+
+export { EvaluationStatus };
