+++ src/evaluation/api/evaluation.controller.ts
@@ -1,18 +1,17 @@
-// Controller that implements evaluation related endpoints
-import { ConversationDocument, ConversationModel } from 'db/models/conversation.model';
-import { SimulationDocument, SimulationModel } from 'db/models/simulation.model';
-import evaluationService from 'evaluation/service/evaluation.service';
-import { RunEvaluationRequest } from 'evaluation/model/request/run-evaluation.request';
-import { Request, Response } from 'express';
-import { INTERNAL_SERVER_ERROR } from 'utils/errors';
+import { ConversationDocument, ConversationModel } from '@db/models/conversation.model';
+import { SimulationDocument, SimulationModel } from '@db/models/simulation.model';
+import { ConversationRepository } from '@db/repositories/conversation.repository';
+import evaluationService from '@evaluation/service/evaluation.service';
+import { RunEvaluationRequest } from '@evaluation/model/request/run-evaluation.request';
+import { INTERNAL_SERVER_ERROR } from '@utils/errors';
+import { EvaluationStatus } from '@enums/evaluation-status.enum';
 import { validationResult } from 'express-validator';
-import { ConversationRepository } from 'db/repositories/conversation.repository';
+import { Request, Response } from 'express';
 import {
   EvaluationResultForConversation,
   EvaluationResultForSimulation,
-  EvaluationStatus,
 } from '@evaluation/model/response/evaluation-result.response';
-import { SimulationRepository } from 'db/repositories/simulation.repository';
+import { SimulationRepository } from '@db/repositories/simulation.repository';
 import { RunEvaluationResponse } from '@evaluation/model/response/run-evaluation.response';
 
 const conversationRepository = new ConversationRepository(ConversationModel);
