+++ src/evaluation/router/evaluation.router.ts
@@ -1,9 +1,9 @@
-// Router that routes incoming requests to relevant controller function
 import express from 'express';
+
 const router = express.Router();
 
-import evaluationController from 'evaluation/api/evaluation.controller';
-import evaluationValidator from 'evaluation/validator/evaluation.validator';
+import evaluationController from '@evaluation/api/evaluation.controller';
+import evaluationValidator from '@evaluation/validator/evaluation.validator';
 
 // region GET //
 router.get(
