+++ src/evaluation/service/metric.service.ts
@@ -1,20 +1,21 @@
-import { MsgSender, MsgTypes } from 'db/enum/enums';
-import { MessageDocument } from 'db/models/message.model';
-import { MetricRepository } from 'db/repositories/metric.repository';
-import { MetricDocument, MetricModel } from 'db/models/metric.model';
-import { MetricNameEnum, metricWeightMap } from 'evaluation/utils/metric.config';
+import { MsgSender } from '@enums/msg-sender.enum';
+import { MsgType } from '@enums/msg-type.enum';
+import { MetricNameEnum } from '@enums/metric-name.enum';
+import { MessageDocument } from '@db/models/message.model';
+import { MetricRepository } from '@db/repositories/metric.repository';
+import { MetricDocument, MetricModel } from '@db/models/metric.model';
+import { ConversationDocument } from '@db/models/conversation.model';
+import { metricWeightMap } from '@evaluation/utils/metric.config';
 import mongoose, { Types } from 'mongoose';
-import { ConversationDocument } from 'db/models/conversation.model';
-import { metricNormalizationFunctions } from 'evaluation/utils/data.normalization';
+import { metricNormalizationFunctions } from '@evaluation/utils/data.normalization';
 
 const metricRepository = new MetricRepository(MetricModel);
 
 /**
  * calculates all metrics
- * @param messages - messages of conversation which should be evaluated
- * @param usedEndpoints - used endpoints of conversation which should be evaluated
- * @throws if values are missing in one of the maps metricCalculationFunctions, metricNormalizationFunctions and metricWeightMap
- * @returns the generated metric documents
+ * @param conversation - Conversation document to be calculated.
+ * @throws Error - If values are missing in one of the maps metricCalculationFunctions, metricNormalizationFunctions and metricWeightMap.
+ * @returns MetricDocument[] -  Generated metric documents for the given conversation.
  */
 async function calculateAllMetrics(conversation: ConversationDocument): Promise<MetricDocument[]> {
   const messages = (await conversation.populate('messages')).messages as MessageDocument[];
@@ -36,10 +37,10 @@ async function calculateAllMetrics(conversation: ConversationDocument): Promise<
 }
 
 /**
- * Calculates the weighted average score
- * @param metrics - metric documents
- * @param useRawValue - whether to calculate average over rawValue or over normalized value
- * @returns weighted average over the scores of all metrics
+ * Calculates the weighted average score using given metric documents.
+ * @param metrics - Metric documents.
+ * @param useRawValue - Whether to calculate average over rawValue or over normalized value.
+ * @returns number -  Weighted average over the scores of all metrics provided in metric documents.
  */
 function calculateWeightedAverage(metrics: MetricDocument[], useRawValue: boolean = false): number {
   return metrics.reduce<number>((accumulator: number, metric: MetricDocument | Types.ObjectId) => {
@@ -49,10 +50,10 @@ function calculateWeightedAverage(metrics: MetricDocument[], useRawValue: boolea
 }
 
 /**
- * Calculates the equal (classical) average score
- * @param metrics - metric documents
- * @param useRawValue - whether to calculate average over rawValue or over normalized value
- * @returns equal overage over the scores of all metrics
+ * Calculates the equal (classical) average score.
+ * @param metrics - Metric documents.
+ * @param useRawValue - Whether to calculate average over rawValue or over normalized value.
+ * @returns number - Equal overage over the scores of all metrics.
  */
 function calculateEqualAverage(metrics: MetricDocument[], useRawValue: boolean = false): number {
   if (metrics.length == 0) return 0;
@@ -65,17 +66,17 @@ function calculateEqualAverage(metrics: MetricDocument[], useRawValue: boolean =
 }
 
 /**
- * Initializes a new metric document in the database
- * @param metricName - name of the metric which is initialized
- * @param score - score (between 0 and 1) that was achieved in the specified metric
- * @param rawValue - not normalized score that was achieved
- * @throws if a value is missing in metricWeightMap
- * @returns the created metric document
+ * Initializes a new metric document in the database.
+ * @param metricName - Name of the metric which is initialized.
+ * @param score - Score (between 0 and 1) that was achieved in the specified metric.
+ * @param rawValue - Not normalized score that was achieved.
+ * @throws Error - If a value is missing in metricWeightMap.
+ * @returns MetricDocument - Created metric document.
  */
-async function initialize(metricName: MetricNameEnum, score: number, rawValue: number) {
+async function initialize(metricName: MetricNameEnum, score: number, rawValue: number): Promise<MetricDocument> {
   const weight = metricWeightMap.get(metricName);
   if (!weight) {
-    throw new Error(`ERROR: metric: missing entry for ${metricName} in metricWeightMap`);
+    throw new Error(`Metric: missing entry for ${metricName} in metricWeightMap`);
   }
   return await metricRepository.create({
     name: metricName,
@@ -95,39 +96,39 @@ const metricCalculationFunctions = new Map<
 ]);
 
 /**
- * Evaluates whether the user's concern was fulfilled or not
- * @param messages - The messages of the conversation
+ * Evaluates whether the user's concern was fulfilled or not.
+ * @param messages - The messages of the conversation.
  * @param _usedEndpoints - (unused) The endpoints that were called during the conversation
- * @returns 1 if the user's concern was fulfilled and 0 if not
+ * @returns number -  If the user's concern was fulfilled and 0 otherwise.
  */
 function goalFulfilled(
   messages: MessageDocument[],
   _usedEndpoints: string[] /* eslint-disable-line @typescript-eslint/no-unused-vars*/,
-) {
-  const result: boolean = messages.length > 0 && messages[messages.length - 1].type == MsgTypes.HANGUP;
+): number {
+  const result: boolean = messages.length > 0 && messages[messages.length - 1].type == MsgType.HANGUP;
   return Number(result);
 }
 
 /**
  * Calculates the number of steps per used endpoint.
- * @param messages - The messages of the conversation
- * @param usedEndpoints - The endpoints that were called during the conversation
- * @returns Number of steps per used endpoint.
+ * @param messages - The messages of the conversation.
+ * @param usedEndpoints - The endpoints that were called during the conversation.
+ * @returns number - Number of steps per used endpoint.
  */
-function countSteps(messages: MessageDocument[], usedEndpoints: string[]) {
+function countSteps(messages: MessageDocument[], usedEndpoints: string[]): number {
   return messages.length / usedEndpoints.length;
 }
 
 /**
  * Calculates the average response time of the agent in a conversation.
- * @param messages - The messages of the conversation
- * @param _usedEndpoints - (unused) The endpoints that were called during the conversation
- * @returns Average response time of the agent in the conversation.
+ * @param messages - The messages of the conversation.
+ * @param _usedEndpoints - (unused) The endpoints that were called during the conversation.
+ * @returns number - Average response time of the agent in the conversation.
  */
 function calculateAverageResponseTime(
   messages: MessageDocument[],
   _usedEndpoints: string[] /* eslint-disable-line @typescript-eslint/no-unused-vars*/,
-) {
+): number {
   if (messages.length < 2) {
     return 0;
   }
@@ -149,8 +150,7 @@ function calculateAverageResponseTime(
     countAgentMessages++;
   }
 
-  const averageResponseTime = totalResponseTimeOfAgent / countAgentMessages;
-  return averageResponseTime;
+  return totalResponseTimeOfAgent / countAgentMessages;
 }
 
 export default { initialize, calculateAllMetrics, calculateWeightedAverage, calculateEqualAverage };
